services/ingest_chunker/requirements.txt
fastapi
uvicorn
pydantic
python-multipart

google-cloud-storage
google-cloud-bigquery

httpx
pypdf
python-docx